// src/types/common.ts
export enum response_status_codes {
  success = 200,
  created = 201,
  no_content = 204,
  permanent_redirect = 301,
  modified = 304,
  temporary_redirect = 307,
  bad_request = 400,
  unauthorized = 401,
  forbidden = 403,
  not_found = 404,
  internal_server_error = 500,
  not_implemented = 501,
}